mysql/requirements.txt
mysql-connector-python
fastapi
uvicorn
python-dotenv